pyproject.toml
[tool.poetry]
name = "morphview"
version = "0.1.0"
description = ""
authors = ["max31ru12 <maxevg72@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
asgiref = "^3.7.2"
django-cors-headers = "^4.3.1"
django-environ = "^0.11.2"
django-debug-toolbar = "^4.3.0"
gunicorn = "^21.2.0"
unidecode = "^1.3.8"
tzdata = "^2024.1"
psycopg2 = "^2.9.9"
redis = "^5.0.3"
django-js-asset = "^2.2.0"
django-ckeditor = "^6.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
